app/requirements.txt
flask==1.1.2
flask-cors==3.0.9
Flask-WTF==0.14.3
passlib==1.7.4
logmatic-python==0.1.7
requests==2.25.1
